Create readers.py

+++ readers.py
@@ -180,7 +180,7 @@ class YT8MFrameFeatureReader(BaseReader):
       num_frames: number of frames in the sequence
     """
     decoded_features = tf.reshape(
-        tf.cast(tf.decode_raw(features, tf.uint8), tf.float32),
+        tf.cast(tf.decode_raw(features, tf.uint8), tf.int16),
         [-1, feature_size])
 
     num_frames = tf.minimum(tf.shape(decoded_features)[0], max_frames)
